Extract AI support power handling to AISupportPowerManager
