+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -127,6 +127,7 @@
     <Compile Include="AI\BaseBuilder.cs" />
     <Compile Include="AI\HackyAI.cs" />
     <Compile Include="AI\AIHarvesterManager.cs" />
+    <Compile Include="AI\AISupportPowerManager.cs" />
     <Compile Include="AI\Squad.cs" />
     <Compile Include="AI\StateMachine.cs" />
     <Compile Include="AI\States\AirStates.cs" />
